+++ src/Symfony/Component/Serializer/Normalizer/UnwrappingDenormalizer.php
@@ -14,15 +14,14 @@ namespace Symfony\Component\Serializer\Normalizer;
 use Symfony\Component\PropertyAccess\PropertyAccess;
 use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 use Symfony\Component\Serializer\Exception\LogicException;
-use Symfony\Component\Serializer\SerializerAwareInterface;
-use Symfony\Component\Serializer\SerializerAwareTrait;
+use Symfony\Component\Serializer\SerializerInterface;
 
 /**
  * @author Eduard Bulava <bulavaeduard@gmail.com>
  */
-final class UnwrappingDenormalizer implements DenormalizerInterface, SerializerAwareInterface
+final class UnwrappingDenormalizer implements DenormalizerInterface, DenormalizerAwareInterface
 {
-    use SerializerAwareTrait;
+    use DenormalizerAwareTrait;
 
     public const UNWRAP_PATH = 'unwrap_path';
 
@@ -51,15 +50,25 @@ final class UnwrappingDenormalizer implements DenormalizerInterface, SerializerA
             $data = $this->propertyAccessor->getValue($data, $propertyPath);
         }
 
-        if (!$this->serializer instanceof DenormalizerInterface) {
-            throw new LogicException('Cannot unwrap path because the injected serializer is not a denormalizer.');
-        }
-
-        return $this->serializer->denormalize($data, $type, $format, $context);
+        return $this->denormalizer->denormalize($data, $type, $format, $context);
     }
 
     public function supportsDenormalization(mixed $data, string $type, string $format = null, array $context = []): bool
     {
         return \array_key_exists(self::UNWRAP_PATH, $context) && !isset($context['unwrapped']);
+    }
+
+    /**
+     * @deprecated Since symfony/serializer 7.1: The "setSerializer()" method is deprecated, use "setDenormalizer()" instead.
+     */
+    public function setSerializer(SerializerInterface $serializer): void
+    {
+        trigger_deprecation('symfony/serializer', '7.1', 'The "%s()" method is deprecated, use "setDenormalizer()" instead.', __METHOD__);
+
+        if (!$serializer instanceof DenormalizerInterface) {
+            throw new LogicException(sprintf('Cannot set denormalizer because the injected serializer does not implement the "%s".', DenormalizerInterface::class));
+        }
+
+        $this->setDenormalizer($serializer);
     }
 }